+++ brockassemble/exceptions.py
@@ -1,0 +1,7 @@
+
+class MissingParameterError(Exception):
+    """
+    An exception which is raised when a function is not given the necessary
+    parameters to execute.
+    """
+    pass